+++ src/helpers/convertMarkdownToJson.js
@@ -1,0 +1,13 @@
+import mdToJson from './md2json';
+
+export default function convertMarkdownToJson(markdown) {
+  let json = {};
+
+  try {
+    json = mdToJson.parse(markdown);
+  } catch (error) {
+    json = { '': { raw: markdown } }
+  }
+
+  return json;
+}
