+++ src/helpers/parseNumber.js
@@ -1,0 +1,9 @@
+export default function parseNumber(number) {
+  if (typeof number === 'string') {
+    if (number.includes('.')) number = number.split('.').slice(0, -1).join('.')
+  }
+
+  const result = typeof number === 'number' || !isNaN(parseInt(number)) ?
+    parseInt(number) : number;
+  return result;
+}
